+++ crates/threshold-signature-server/src/helpers/tests.rs
@@ -25,6 +25,7 @@ use crate::{
     chain_api::{
         entropy::{
             self,
+            runtime_types::entropy_runtime::RuntimeCall,
             runtime_types::pallet_staking_extension::pallet::{JumpStartStatus, ServerInfo},
         },
         EntropyConfig,
@@ -43,6 +44,7 @@ use entropy_protocol::PartyId;
 #[cfg(test)]
 use entropy_shared::EncodedVerifyingKey;
 use entropy_shared::NETWORK_PARENT_KEY;
+use sp_keyring::AccountKeyring;
 use std::{fmt, net::SocketAddr, str, time::Duration};
 use subxt::{
     backend::legacy::LegacyRpcMethods, ext::sp_core::sr25519, tx::PairSigner,
@@ -353,4 +355,19 @@ pub fn get_port(server_info: &ServerInfo<SubxtAccountId32>) -> u32 {
     let socket_address: SocketAddr =
         str::from_utf8(&server_info.endpoint).unwrap().parse().unwrap();
     socket_address.port().into()
+}
+
+/// Calls set storage from sudo for testing, allows use to manipulate chain storage.
+pub async fn call_set_storage(
+    api: &OnlineClient<EntropyConfig>,
+    rpc: &LegacyRpcMethods<EntropyConfig>,
+    call: RuntimeCall,
+) {
+    let set_storage = entropy::tx().sudo().sudo(call);
+    let alice = AccountKeyring::Alice;
+
+    let signature_request_pair_signer =
+        PairSigner::<EntropyConfig, sp_core::sr25519::Pair>::new(alice.into());
+
+    submit_transaction(api, rpc, &signature_request_pair_signer, &set_storage, None).await.unwrap();
 }