+++ node/cli/src/chain_spec/testnet.rs
@@ -427,7 +427,6 @@ pub fn testnet_genesis_config(
                 })
                 .collect::<Vec<_>>(),
             proactive_refresh_data: (vec![], vec![]),
-            mock_signer_rotate: (false, vec![], vec![]),
             jump_started_signers: None,
         },
         "elections": ElectionsConfig {