+++ pallets/staking/src/benchmarking.rs
@@ -16,7 +16,6 @@
 //! Benchmarking setup for pallet-propgation
 #![allow(unused_imports)]
 use super::*;
-use crate::pck::{signing_key_from_seed, MOCK_PCK_DERIVED_FROM_NULL_ARRAY};
 #[allow(unused_imports)]
 use crate::Pallet as Staking;
 use entropy_shared::{AttestationHandler, QuoteContext, MAX_SIGNERS};
@@ -34,8 +33,14 @@ use pallet_staking::{
     Event as FrameStakingEvent, MaxNominationsOf, MaxValidatorsCount, Nominations,
     Pallet as FrameStaking, RewardDestination, ValidatorPrefs,
 };
+use rand::{rngs::StdRng, SeedableRng};
 use sp_std::{vec, vec::Vec};
+use tdx_quote::SigningKey;
 
+const MOCK_PCK_DERIVED_FROM_NULL_ARRAY: [u8; 33] = [
+    3, 237, 193, 27, 177, 204, 234, 67, 54, 141, 157, 13, 62, 87, 113, 224, 4, 121, 206, 251, 190,
+    151, 134, 87, 68, 46, 37, 163, 127, 97, 252, 174, 108,
+];
 const NULL_ARR: [u8; 32] = [0; 32];
 const SEED: u32 = 0;
 
@@ -89,6 +94,7 @@ fn prepare_attestation_for_validate<T: Config>(
     threshold: T::AccountId,
     x25519_public_key: [u8; 32],
     endpoint: Vec<u8>,
+    quote_context: QuoteContext,
 ) -> (Vec<u8>, JoiningServerInfo<T::AccountId>) {
     let nonce = NULL_ARR;
     let quote = {
@@ -105,25 +111,26 @@ fn prepare_attestation_for_validate<T: Config>(
             &threshold,
             x25519_public_key,
             nonce,
-            QuoteContext::Validate,
+            quote_context,
         );
-
-        tdx_quote::Quote::mock(attestation_key.clone(), pck, input_data.0).as_bytes().to_vec()
+        let pck_encoded = tdx_quote::encode_verifying_key(pck.verifying_key()).unwrap();
+        tdx_quote::Quote::mock(attestation_key.clone(), pck, input_data.0, pck_encoded.to_vec())
+            .as_bytes()
+            .to_vec()
     };
 
-    let joining_server_info = JoiningServerInfo {
-        tss_account: threshold.clone(),
-        x25519_public_key,
-        endpoint,
-        // Since we are using the mock PckCertChainVerifier, this needs to be the same seed for
-        // generating the PCK as we used to sign the quote above
-        pck_certificate_chain: vec![NULL_ARR.to_vec()],
-    };
+    let joining_server_info =
+        JoiningServerInfo { tss_account: threshold.clone(), x25519_public_key, endpoint };
 
     // We need to tell the attestation handler that we want a quote. This will let the system to
     // know to expect one back when we call `validate()`.
     T::AttestationHandler::request_quote(&threshold, nonce);
     (quote, joining_server_info)
+}
+
+fn signing_key_from_seed(input: [u8; 32]) -> SigningKey {
+    let mut pck_seeder = StdRng::from_seed(input);
+    SigningKey::random(&mut pck_seeder)
 }
 
 fn prep_bond_and_validate<T: Config>(
@@ -147,8 +154,12 @@ fn prep_bond_and_validate<T: Config>(
 
     if validate_also {
         let endpoint = b"http://localhost:3001".to_vec();
-        let (quote, joining_server_info) =
-            prepare_attestation_for_validate::<T>(threshold, x25519_public_key, endpoint);
+        let (quote, joining_server_info) = prepare_attestation_for_validate::<T>(
+            threshold,
+            x25519_public_key,
+            endpoint,
+            QuoteContext::Validate,
+        );
 
         assert_ok!(<Staking<T>>::validate(
             RawOrigin::Signed(bonder.clone()).into(),
@@ -198,6 +209,7 @@ benchmarks! {
         threshold,
         x25519_public_key,
         endpoint.clone().to_vec(),
+        QuoteContext::ChangeEndpoint,
     )
     .0;
   }:  _(RawOrigin::Signed(bonder.clone()), endpoint.to_vec(), quote)
@@ -239,9 +251,8 @@ benchmarks! {
         new_threshold.clone(),
         x25519_public_key,
         endpoint.clone().to_vec(),
+        QuoteContext::ChangeThresholdAccounts,
     );
-
-    let pck_certificate_chain = joining_server_info.pck_certificate_chain;
 
     let signers = vec![validator_id_signers.clone(); s as usize];
     Signers::<T>::put(signers.clone());
@@ -249,7 +260,6 @@ benchmarks! {
         RawOrigin::Signed(_bonder.clone()),
         new_threshold.clone(),
         x25519_public_key.clone(),
-        pck_certificate_chain,
         quote
     )
   verify {
@@ -388,7 +398,7 @@ benchmarks! {
     );
 
     let (quote, joining_server_info) =
-        prepare_attestation_for_validate::<T>(threshold_account.clone(), x25519_public_key, endpoint.clone());
+        prepare_attestation_for_validate::<T>(threshold_account.clone(), x25519_public_key, endpoint.clone(), QuoteContext::Validate);
   }:  _(RawOrigin::Signed(bonder.clone()), ValidatorPrefs::default(), joining_server_info, quote)
   verify {
     assert_last_event::<T>(