+++ pallets/staking/src/tests.rs
@@ -14,8 +14,8 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 use crate::{
-    mock::*, pck::MOCK_PCK_DERIVED_FROM_NULL_ARRAY, tests::RuntimeEvent, Error, JoiningServerInfo,
-    NextSignerInfo, NextSigners, ServerInfo, Signers,
+    mock::*, tests::RuntimeEvent, Error, JoiningServerInfo, NextSignerInfo, NextSigners,
+    ServerInfo, Signers,
 };
 use codec::Encode;
 use frame_support::{assert_noop, assert_ok};
@@ -61,12 +61,8 @@ fn it_takes_in_an_endpoint() {
             pallet_staking::RewardDestination::Account(1),
         ));
 
-        let joining_server_info = JoiningServerInfo {
-            tss_account: 3,
-            x25519_public_key: NULL_ARR,
-            endpoint: vec![20],
-            pck_certificate_chain: vec![[0u8; 32].to_vec()],
-        };
+        let joining_server_info =
+            JoiningServerInfo { tss_account: 3, x25519_public_key: NULL_ARR, endpoint: vec![20] };
         assert_ok!(Staking::validate(
             RuntimeOrigin::signed(1),
             pallet_staking::ValidatorPrefs::default(),
@@ -83,7 +79,6 @@ fn it_takes_in_an_endpoint() {
             tss_account: 3,
             x25519_public_key: NULL_ARR,
             endpoint: [20; (crate::tests::MaxEndpointLength::get() + 1) as usize].to_vec(),
-            pck_certificate_chain: vec![[0u8; 32].to_vec()],
         };
         assert_noop!(
             Staking::validate(
@@ -99,7 +94,6 @@ fn it_takes_in_an_endpoint() {
             tss_account: 5,
             x25519_public_key: NULL_ARR,
             endpoint: vec![20, 20],
-            pck_certificate_chain: vec![[0u8; 32].to_vec()],
         };
         assert_noop!(
             Staking::validate(
@@ -122,12 +116,8 @@ fn it_will_not_allow_validator_to_use_existing_tss_account() {
             pallet_staking::RewardDestination::Account(1),
         ));
 
-        let joining_server_info = JoiningServerInfo {
-            tss_account: 3,
-            x25519_public_key: NULL_ARR,
-            endpoint: vec![20],
-            pck_certificate_chain: vec![[0u8; 32].to_vec()],
-        };
+        let joining_server_info =
+            JoiningServerInfo { tss_account: 3, x25519_public_key: NULL_ARR, endpoint: vec![20] };
         assert_ok!(Staking::validate(
             RuntimeOrigin::signed(1),
             pallet_staking::ValidatorPrefs::default(),
@@ -168,7 +158,6 @@ fn it_changes_endpoint() {
             tss_account: 3,
             x25519_public_key: NULL_ARR,
             endpoint: endpoint.clone(),
-            pck_certificate_chain: vec![[0u8; 32].to_vec()],
         };
         assert_ok!(Staking::validate(
             RuntimeOrigin::signed(1),
@@ -206,7 +195,6 @@ fn it_doesnt_change_endpoint_with_invalid_quote() {
             tss_account: 3,
             x25519_public_key: NULL_ARR,
             endpoint: endpoint.clone(),
-            pck_certificate_chain: vec![[0u8; 32].to_vec()],
         };
 
         assert_ok!(Staking::validate(
@@ -218,7 +206,7 @@ fn it_doesnt_change_endpoint_with_invalid_quote() {
 
         assert_noop!(
             Staking::change_endpoint(RuntimeOrigin::signed(1), endpoint, INVALID_QUOTE.to_vec()),
-            Error::<Test>::FailedAttestationCheck
+            Error::<Test>::BadQuote
         );
     })
 }
@@ -232,13 +220,8 @@ fn it_changes_threshold_account() {
             pallet_staking::RewardDestination::Account(1),
         ));
 
-        let pck_certificate_chain = vec![vec![0u8; 32]];
-        let joining_server_info = JoiningServerInfo {
-            tss_account: 3,
-            x25519_public_key: NULL_ARR,
-            endpoint: vec![20],
-            pck_certificate_chain: pck_certificate_chain.clone(),
-        };
+        let joining_server_info =
+            JoiningServerInfo { tss_account: 3, x25519_public_key: NULL_ARR, endpoint: vec![20] };
         assert_ok!(Staking::validate(
             RuntimeOrigin::signed(1),
             pallet_staking::ValidatorPrefs::default(),
@@ -250,7 +233,6 @@ fn it_changes_threshold_account() {
             RuntimeOrigin::signed(1),
             4,
             NULL_ARR,
-            pck_certificate_chain.clone(),
             VALID_QUOTE.to_vec()
         ));
         assert_eq!(Staking::threshold_server(1).unwrap().tss_account, 4);
@@ -261,7 +243,6 @@ fn it_changes_threshold_account() {
                 RuntimeOrigin::signed(4),
                 5,
                 NULL_ARR,
-                pck_certificate_chain.clone(),
                 VALID_QUOTE.to_vec()
             ),
             Error::<Test>::NotController
@@ -274,12 +255,8 @@ fn it_changes_threshold_account() {
             pallet_staking::RewardDestination::Account(2),
         ));
 
-        let joining_server_info = JoiningServerInfo {
-            tss_account: 5,
-            x25519_public_key: NULL_ARR,
-            endpoint: vec![20],
-            pck_certificate_chain: pck_certificate_chain.clone(),
-        };
+        let joining_server_info =
+            JoiningServerInfo { tss_account: 5, x25519_public_key: NULL_ARR, endpoint: vec![20] };
         assert_ok!(Staking::validate(
             RuntimeOrigin::signed(2),
             pallet_staking::ValidatorPrefs::default(),
@@ -292,7 +269,6 @@ fn it_changes_threshold_account() {
                 RuntimeOrigin::signed(1),
                 5,
                 NULL_ARR,
-                pck_certificate_chain.clone(),
                 VALID_QUOTE.to_vec()
             ),
             Error::<Test>::TssAccountAlreadyExists
@@ -304,7 +280,6 @@ fn it_changes_threshold_account() {
                 RuntimeOrigin::signed(1),
                 9,
                 NULL_ARR,
-                pck_certificate_chain.clone(),
                 VALID_QUOTE.to_vec()
             ),
             Error::<Test>::NoChangingThresholdAccountWhenSigner
@@ -321,13 +296,8 @@ fn it_doesnt_allow_changing_threshold_account_with_invalid_quote() {
             pallet_staking::RewardDestination::Account(1),
         ));
 
-        let pck_certificate_chain = vec![[0u8; 32].to_vec()];
-        let joining_server_info = JoiningServerInfo {
-            tss_account: 3,
-            x25519_public_key: NULL_ARR,
-            endpoint: vec![20],
-            pck_certificate_chain: pck_certificate_chain.clone(),
-        };
+        let joining_server_info =
+            JoiningServerInfo { tss_account: 3, x25519_public_key: NULL_ARR, endpoint: vec![20] };
         assert_ok!(Staking::validate(
             RuntimeOrigin::signed(1),
             pallet_staking::ValidatorPrefs::default(),
@@ -340,10 +310,9 @@ fn it_doesnt_allow_changing_threshold_account_with_invalid_quote() {
                 RuntimeOrigin::signed(1),
                 4,
                 NULL_ARR,
-                pck_certificate_chain.clone(),
                 INVALID_QUOTE.to_vec()
             ),
-            Error::<Test>::FailedAttestationCheck
+            Error::<Test>::BadQuote
         );
     })
 }
@@ -357,13 +326,8 @@ fn it_will_not_allow_existing_tss_account_when_changing_threshold_account() {
             pallet_staking::RewardDestination::Account(1),
         ));
 
-        let pck_certificate_chain = vec![[0u8; 32].to_vec()];
-        let joining_server_info = JoiningServerInfo {
-            tss_account: 3,
-            x25519_public_key: NULL_ARR,
-            endpoint: vec![20],
-            pck_certificate_chain: pck_certificate_chain.clone(),
-        };
+        let joining_server_info =
+            JoiningServerInfo { tss_account: 3, x25519_public_key: NULL_ARR, endpoint: vec![20] };
         assert_ok!(Staking::validate(
             RuntimeOrigin::signed(1),
             pallet_staking::ValidatorPrefs::default(),
@@ -378,12 +342,8 @@ fn it_will_not_allow_existing_tss_account_when_changing_threshold_account() {
             pallet_staking::RewardDestination::Account(2),
         ));
 
-        let joining_server_info = JoiningServerInfo {
-            tss_account: 5,
-            x25519_public_key: NULL_ARR,
-            endpoint: vec![20],
-            pck_certificate_chain: pck_certificate_chain.clone(),
-        };
+        let joining_server_info =
+            JoiningServerInfo { tss_account: 5, x25519_public_key: NULL_ARR, endpoint: vec![20] };
         assert_ok!(Staking::validate(
             RuntimeOrigin::signed(2),
             pallet_staking::ValidatorPrefs::default(),
@@ -396,7 +356,6 @@ fn it_will_not_allow_existing_tss_account_when_changing_threshold_account() {
                 RuntimeOrigin::signed(1),
                 5,
                 NULL_ARR,
-                pck_certificate_chain.clone(),
                 VALID_QUOTE.to_vec()
             ),
             Error::<Test>::TssAccountAlreadyExists
@@ -416,12 +375,8 @@ fn it_deletes_when_no_bond_left() {
             pallet_staking::RewardDestination::Account(1),
         ));
 
-        let joining_server_info = JoiningServerInfo {
-            tss_account: 3,
-            x25519_public_key: NULL_ARR,
-            endpoint: vec![20],
-            pck_certificate_chain: vec![[0u8; 32].to_vec()],
-        };
+        let joining_server_info =
+            JoiningServerInfo { tss_account: 3, x25519_public_key: NULL_ARR, endpoint: vec![20] };
         assert_ok!(Staking::validate(
             RuntimeOrigin::signed(2),
             pallet_staking::ValidatorPrefs::default(),
@@ -555,11 +510,7 @@ fn it_tests_new_session_handler() {
         assert_ok!(Staking::new_session_handler(&[1, 5, 6]));
         // takes signers original (5,6) pops off one and adds in new validator
         assert_eq!(Staking::next_signers().unwrap().next_signers, vec![6, 1]);
-        assert_eq!(
-            Staking::reshare_data().block_number,
-            101,
-            "Check reshare block start at 100 + 1"
-        );
+        assert_eq!(Staking::reshare_data().block_number, 99, "Check reshare block start at 99 + 1");
         assert_eq!(
             Staking::reshare_data().new_signers,
             vec![1u64.encode()],
@@ -569,12 +520,6 @@ fn it_tests_new_session_handler() {
             Staking::jump_start_progress().parent_key_threshold,
             2,
             "parent key threhsold updated"
-        );
-
-        assert_eq!(
-            Staking::reshare_data().block_number,
-            101,
-            "Check reshare block start at 100 + 1"
         );
 
         assert_ok!(Staking::new_session_handler(&[6, 5, 3]));
@@ -702,12 +647,8 @@ fn it_requires_attestation_before_validate_is_succesful() {
             pallet_staking::RewardDestination::Account(alice),
         ));
 
-        let joining_server_info = JoiningServerInfo {
-            tss_account: bob,
-            x25519_public_key: NULL_ARR,
-            endpoint: vec![20],
-            pck_certificate_chain: vec![[0u8; 32].to_vec()],
-        };
+        let joining_server_info =
+            JoiningServerInfo { tss_account: bob, x25519_public_key: NULL_ARR, endpoint: vec![20] };
 
         // First we test that an invalid attestation doesn't allow us to submit our candidacy.
         assert_noop!(
@@ -717,7 +658,7 @@ fn it_requires_attestation_before_validate_is_succesful() {
                 joining_server_info.clone(),
                 INVALID_QUOTE.to_vec(),
             ),
-            Error::<Test>::FailedAttestationCheck
+            Error::<Test>::BadQuote
         );
 
         assert_eq!(Staking::threshold_server(bob), None);
@@ -735,10 +676,7 @@ fn it_requires_attestation_before_validate_is_succesful() {
             tss_account: joining_server_info.tss_account,
             x25519_public_key: joining_server_info.x25519_public_key,
             endpoint: joining_server_info.endpoint,
-            provisioning_certification_key: MOCK_PCK_DERIVED_FROM_NULL_ARRAY
-                .to_vec()
-                .try_into()
-                .unwrap(),
+            provisioning_certification_key: [0; 33].to_vec().try_into().unwrap(),
         };
         assert_eq!(Staking::threshold_to_stash(bob), Some(alice));
         assert_eq!(Staking::threshold_server(alice), Some(server_info));